refactor to have User superclass, add Language table, add TeachingStyle enum
