+++ backend/src/main/java/com/triolingo/entity/User.java
@@ -1,0 +1,22 @@
+package com.triolingo.entity;
+
+import jakarta.persistence.*;
+import lombok.Data;
+import javax.validation.constraints.Email;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Pattern;
+
+@MappedSuperclass
+@Data
+public abstract class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @NotNull
+    private String fullName;
+    @NotNull
+    @Email(regexp = "[a-z0-9._%+-]+@[a-z0-9.-]+\\.[a-z]{2,3}", flags = Pattern.Flag.CASE_INSENSITIVE)
+    private String email;
+    @NotNull
+    private String password;
+}
