+++ backend/src/main/java/com/triolingo/entity/teacher/TeachingStyle.java
@@ -1,0 +1,7 @@
+package com.triolingo.entity.teacher;
+
+public enum TeachingStyle {
+    GROUP,
+    INDIVIDUAL,
+    FLEXIBLE
+}
